Settings: Advanced low battery indicator options (2/2) Based on a idea I saw on Liquid and from CR3C0 (thanks guys) which we modified to an extended version
This commit introduces following new SlimBean modes to indicate low battery

     * 0 = show dialog + play sound (default)
     * 1 = fire notification + play sound
     * 2 = show dialog only
     * 3 = fire notification only
     * 4 = play sound only
     * 5 = none

Due that notification is a unique SlimBean feature some words on it. Notification
behaviour is the same as Pop-up. User can dismiss it manually or it is
dismissed automatically if charger is connected. Also you user get a
status bar ticker message about the notification and notification is always
on first place of list. And last if vibration is enabled
notification vibrates.

cheers

Change-Id: Ia8fa8c420e6ef334266a7734fb080026a8396340

Conflicts:
	res/values/cm_arrays.xml
	res/values/cm_strings.xml
	res/xml/display_settings.xml
	src/com/android/settings/DisplaySettings.java